Generator

+++ Generator.cpp
@@ -10,7 +10,7 @@ public:
     void test(void) {
         long int *z = new long(INT_MAX);
         // time_t *t = time_t(z);
-        srand(time(new long(INT_MIN))+time(new long(INT_MAX)));
+        srand(time(new long(LONG_MIN))+time(new long(LONG_MAX)));
         int n = 10;
         std::vector<int> arr(n);
         _for(i, n) arr[i] = rand();
